refactor: reduce complexity of action types - place and view bid as same action type
